refactor: move `knotspace`, `functions` and `curves` to submodule
