+++ src/pynurbs/__init__.py
@@ -1,8 +1,6 @@
 from .advanced import Intersection, Projection
 from .calculus import Derivate, Integrate
-from .curves import Curve
-from .functions import Function
-from .knotspace import GeneratorKnotVector, KnotVector
+from .responsive import Curve, Function, GeneratorKnotVector, KnotVector
 
 __version__ = "1.1.0"
