+++ src/pynurbs/advanced.py
@@ -7,9 +7,9 @@ from typing import Any, Tuple
 
 import numpy as np
 
-from pynurbs import heavy
-from pynurbs.calculus import Derivate
-from pynurbs.curves import Curve
+from . import heavy
+from .calculus import Derivate
+from .responsive.curves import Curve
 
 
 class Projection:
