+++ src/pynurbs/responsive/knotspace.py
@@ -12,8 +12,8 @@ from typing import Optional, Tuple, Union
 
 import numpy as np
 
-from .core import ImmutableKnotVector
-from .operations.knotvector import (
+from ..core import ImmutableKnotVector
+from ..operations.knotvector import (
     decrease_degree,
     increase_degree,
     insert_knots,
