+++ tests/responsive/test_beziercurve.py
@@ -6,8 +6,8 @@ from fractions import Fraction as frac
 import numpy as np
 import pytest
 
-from pynurbs.curves import Curve
-from pynurbs.knotspace import GeneratorKnotVector, KnotVector
+from pynurbs.responsive.curves import Curve
+from pynurbs.responsive.knotspace import GeneratorKnotVector, KnotVector
 
 
 @pytest.mark.order(4)
