+++ tests/responsive/test_splinecurve.py
@@ -3,8 +3,8 @@ from copy import copy
 import numpy as np
 import pytest
 
-from pynurbs.curves import Curve
-from pynurbs.knotspace import GeneratorKnotVector, KnotVector
+from pynurbs.responsive.curves import Curve
+from pynurbs.responsive.knotspace import GeneratorKnotVector, KnotVector
 
 
 @pytest.mark.order(5)
