+++ tests/test_fitting.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
-from pynurbs.curves import Curve
-from pynurbs.knotspace import GeneratorKnotVector
+from pynurbs.responsive.curves import Curve
+from pynurbs.responsive.knotspace import GeneratorKnotVector
 
 
 @pytest.mark.order(7)
